Move mutation to gem
