+++ Gemfile
@@ -5,6 +5,7 @@ require "pp"
 gemspec
 
 gem "activesupport", "6.1.4.1"
+gem "firebase_id_token", "2.4.0"
 gem "foreman", "0.87.2"
 gem "google-cloud-pubsub", "2.8.1"
 gem "graphql", "1.12.19"
