+++ lib/souls/app/index.rb
@@ -1,3 +1,4 @@
+require_relative "graphql/souls_mutation"
 require_relative "graphql/souls_query"
 require_relative "utils/souls_logger"
