+++ lib/souls/index.rb
@@ -1,6 +1,7 @@
 require_relative "./version"
 require "thor"
 require "graphql"
+require "firebase_id_token"
 require_relative "./cli/index"
 require_relative "./utils/index"
 require_relative "./app/index"
